29 implement alarm rules for ntenums (#30)
* Rules into separate files, addition of new alarm NTEnum non-standard rule

* Add handler constructors interface to SharedNT

* Changes to NTEnum tests and example because of new alarm NTEnum rule
