+++ p4pillon/rules/rules.py
@@ -9,21 +9,18 @@ from __future__ import annotations
 
 import itertools
 import logging
-import operator
-import time
 from abc import ABC, abstractmethod
 from copy import deepcopy
 from enum import IntEnum, auto
 from functools import wraps
 from typing import Any  # Hack to type hint number types
-from typing import SupportsFloat as Numeric
 
 from p4p import Type, Value
 from p4p.server import ServerOperation
-from p4p.server.raw import ServOpWrap, SharedPV
+from p4p.server.raw import ServOpWrap
 
-from p4pillon.definitions import AlarmSeverity
-from p4pillon.utils import overwrite_marked, time_in_seconds_and_nanoseconds
+from p4pillon.server.raw import SharedPV
+from p4pillon.utils import overwrite_marked
 
 logger = logging.getLogger(__name__)
 
@@ -279,268 +276,6 @@ class BaseArrayRule(BaseRule, ABC):
     """
     Rule to be applied to NTScalarArrays
     """
-
-
-class ReadOnlyRule(BaseScalarRule):
-    """A rule which rejects all attempts to put values"""
-
-    @property
-    def _name(self) -> str:
-        return "read_only"
-
-    @property
-    def _fields(self) -> list[str]:
-        return []
-
-    def put_rule(self, pv: SharedPV, op: ServerOperation) -> RulesFlow:
-        return RulesFlow(RulesFlow.ABORT).set_errormsg("read-only")
-
-
-class AlarmRule(BaseRule):
-    """
-    This class exists only to allow the alarm field, i.e. severity and
-    message to be made read-only for put operations
-    """
-
-    @property
-    def _name(self) -> str:
-        return "alarm"
-
-    @property
-    def _fields(self) -> list[str]:
-        return ["alarm"]
-
-
-class TimestampRule(BaseRule):
-    """Set current timestamp unless provided with an alternative value"""
-
-    @property
-    def _name(self) -> str:
-        return "timestamp"
-
-    @property
-    def _fields(self) -> list[str]:
-        return ["timeStamp"]
-
-    def is_applicable(self, newpvstate: Value) -> bool:
-        """
-        Override the base class's rule because timeStamp changes are triggered
-        by changes to any field and not just to the timeStamp field
-        """
-        # If nothing at all has changed then don't update the timeStamp
-        # TODO: Check if this is expected behaviour for Normative Types
-        if not newpvstate.changedSet():
-            return False
-
-        # Check if there is a timeStamp field to update!
-        if "timeStamp" not in newpvstate.keys():
-            return False
-
-        return True
-
-    @check_applicable_init
-    def init_rule(self, newpvstate: Value) -> RulesFlow:
-        """Update the timeStamp of a PV"""
-
-        seconds, nanoseconds = time_in_seconds_and_nanoseconds(time.time())
-        # TODO: there's a bug in the _wrap which means that timestamps are always marked as changed
-        #       Fix this when that bug is fixed.
-        # if "timeStamp.secondsPastEpoch" not in newpvstate.changedSet():
-        if True:
-            logger.debug("using secondsPastEpoch from time.time()")
-            newpvstate["timeStamp.secondsPastEpoch"] = seconds
-        # if "timeStamp.nanoseconds" not in newpvstate.changedSet():
-        if True:
-            newpvstate["timeStamp.nanoseconds"] = nanoseconds
-            logger.debug("using nanoseconds from time.time()")
-
-        return RulesFlow.CONTINUE
-
-
-class ControlRule(BaseScalarRule):
-    """
-    Apply rules implied by Normative Type control field.
-    These include a minimum value change (control.minStep) and upper
-    and lower limits for values (control.limitHigh and control.limitLow)
-    """
-
-    @property
-    def _name(self) -> str:
-        return "control"
-
-    @property
-    def _fields(self) -> list[str]:
-        return ["control"]
-
-    @check_applicable_init
-    def init_rule(self, newpvstate: Value) -> RulesFlow:
-        """Check whether a value should be clipped by the control limits
-
-        NOTE: newpvstate from a put is a combination of the old and new state
-
-        Returns None if no change should be made and the value is valid
-
-        TODO: see if this can be separated out into a function like the
-        min_step_violated to work better with arrays
-
-        """
-        logger.debug("Evaluating control.init rule")
-
-        # Check lower and upper control limits
-        if newpvstate["value"] < newpvstate["control.limitLow"]:
-            newpvstate["value"] = newpvstate["control.limitLow"]
-            logger.debug(
-                "Lower control limit exceeded, changing value to %s",
-                newpvstate["value"],
-            )
-            return RulesFlow.CONTINUE
-
-        if newpvstate["value"] > newpvstate["control.limitHigh"]:
-            newpvstate["value"] = newpvstate["control.limitHigh"]
-            logger.debug(
-                "Upper control limit exceeded, changing value to %s",
-                newpvstate["value"],
-            )
-            return RulesFlow.CONTINUE
-
-        return RulesFlow.CONTINUE
-
-    @check_applicable_post
-    def post_rule(self, oldpvstate: Value, newpvstate: Value) -> RulesFlow:
-        logger.debug("Evaluating control.post rule")
-        # Check minimum step first - if the check for the minimum step fails then we continue
-        # and ignore the actual evaluation of the limits
-        if __class__.min_step_violated(
-            newpvstate["value"],
-            oldpvstate["value"],
-            newpvstate["control.minStep"],
-        ):
-            logger.debug("<minStep")
-            newpvstate["value"] = oldpvstate["value"]
-
-        # if the min step isn't violated, we continue and evaluate the limits themselves
-        # on the value
-        return self.init_rule(newpvstate)
-
-    @classmethod
-    def min_step_violated(cls, new_val, old_val, min_step) -> Numeric:
-        """Check whether the new value is too small to pass a minStep threshold"""
-        if old_val is None or min_step is None:
-            return False
-
-        return abs(new_val - old_val) < min_step
-
-
-class ValueAlarmRule(BaseGatherableRule):
-    """
-    Rule to check whether valueAlarm limits have been triggered, changing
-    alarm.severity and alarm.message appropriately.
-
-    TODO: Implement hysteresis
-    """
-
-    @property
-    def _name(self) -> str:
-        return "valueAlarm"
-
-    @property
-    def _fields(self) -> list[str]:
-        return ["alarm", "valueAlarm"]
-
-    @check_applicable_init
-    def init_rule(self, newpvstate: Value) -> RulesFlow:
-        """Evaluate alarm value limits"""
-        # TODO: Apply the rule for hysteresis. Unfortunately I don't understand the
-        # explanation in the Normative Types specification...
-        logger.debug("Evaluating %s.init_rule", self._name)
-
-        # Check if valueAlarms are present and active!
-        if not newpvstate["valueAlarm.active"]:
-            # TODO: This is wrong! If valueAlarm was active and then made inactive
-            #       the alarm will not be cleared
-            logger.debug("\tvalueAlarm not active")
-            return RulesFlow.CONTINUE
-
-        try:
-            # The order of these tests is defined in the Normative Types document
-            if self.__alarm_state_check(newpvstate, "highAlarm"):
-                return RulesFlow.CONTINUE
-            if self.__alarm_state_check(newpvstate, "lowAlarm"):
-                return RulesFlow.CONTINUE
-            if self.__alarm_state_check(newpvstate, "highWarning"):
-                return RulesFlow.CONTINUE
-            if self.__alarm_state_check(newpvstate, "lowWarning"):
-                return RulesFlow.CONTINUE
-        except SyntaxError:
-            # TODO: Need more specific error than SyntaxError and to decide
-            # if continue is the correct behaviour
-            return RulesFlow.CONTINUE
-
-        # If we made it here then there are no alarms or warnings and we need to indicate that
-        # possibly by resetting any existing ones
-        alarms_changed = False
-        if newpvstate["alarm.severity"]:
-            newpvstate["alarm.severity"] = 0
-            alarms_changed = True
-        if newpvstate["alarm.message"]:
-            newpvstate["alarm.message"] = ""
-            alarms_changed = True
-
-        if alarms_changed:
-            logger.debug(
-                "Setting to severity %i with message '%s'",
-                newpvstate["alarm.severity"],
-                newpvstate["alarm.message"],
-            )
-        else:
-            logger.debug("Made no automatic changes to alarm state.")
-
-        return RulesFlow.CONTINUE
-
-    @classmethod
-    def __alarm_state_check(cls, pvstate: Value, alarm_type: str, op=None) -> bool:
-        """Check whether the PV should be in an alarm state"""
-        if not op:
-            if alarm_type.startswith("low"):
-                op = operator.le
-            elif alarm_type.startswith("high"):
-                op = operator.ge
-            else:
-                raise SyntaxError(f"CheckAlarms/alarmStateCheck: do not know how to handle {alarm_type}")
-
-        severity = pvstate[f"valueAlarm.{alarm_type}Severity"]
-        if op(pvstate["value"], pvstate[f"valueAlarm.{alarm_type}Limit"]) and severity:
-            pvstate["alarm.severity"] = severity
-
-            # TODO: Understand this commented out code!
-            #       I think it's to handle the case of an INVALID alarm or manually
-            #       set alarm message?
-            # if not pvstate.changed("alarm.message"):
-            #     pvstate["alarm.message"] = alarm_type
-            pvstate["alarm.message"] = alarm_type
-
-            logger.debug(
-                "Setting to severity %i with message '%s'",
-                severity,
-                pvstate["alarm.message"],
-            )
-
-            return True
-
-        return False
-
-    def gather_init(self, gathered_value: Value) -> None:
-        if (
-            not gathered_value.changed("alarm.severity")
-            or gathered_value["alarm.severity"] != AlarmSeverity.INVALID_ALARM
-        ):
-            gathered_value["alarm.severity"] = AlarmSeverity.NO_ALARM
-            gathered_value["alarm.message"] = ""
-
-    def gather(self, scalar_value: Value, gathered_value: Value) -> None:
-        if scalar_value["alarm.severity"] > gathered_value["alarm.severity"]:
-            gathered_value["alarm.severity"] = scalar_value["alarm.severity"]
-            gathered_value["alarm.message"] = scalar_value["alarm.message"]
 
 
 class ScalarToArrayWrapperRule(BaseArrayRule):
