+++ p4pillon/thread/sharednt.py
@@ -6,11 +6,13 @@ from __future__ import annotations
 
 import logging
 from collections import OrderedDict
+from typing import Any
 
 from p4pillon.composite_handler import CompositeHandler
 from p4pillon.nt import NTBase, NTEnum, NTScalar
 from p4pillon.nthandlers import ComposeableRulesHandler
 from p4pillon.rules import (
+    AlarmNTEnumRule,
     AlarmRule,
     ControlRule,
     ScalarToArrayWrapperRule,
@@ -33,6 +35,7 @@ class SharedNT(SharedPV):
         self,
         auth_handlers: OrderedDict[str, Handler] | None = None,
         user_handlers: OrderedDict[str, Handler] | None = None,
+        handler_constructors: dict[str, Any] | None = None,
         **kws,
     ):
         # Check if there is a handler specified in the kws, and if not override it
@@ -67,6 +70,12 @@ class SharedNT(SharedPV):
                     else:
                         raise TypeError(f"Unrecognised NT type: {nttype_str}")
                 case NTEnum():
+                    handler["alarm"] = ComposeableRulesHandler(AlarmRule())
+
+                    alarm_ntenum_constructor = None
+                    if handler_constructors:
+                        alarm_ntenum_constructor = handler_constructors.get("alarmNTEnum", None)
+                    handler["alarmNTEnum"] = ComposeableRulesHandler(AlarmNTEnumRule(alarm_ntenum_constructor))
                     handler["timestamp"] = ComposeableRulesHandler(TimestampRule())
                 case _:
                     raise NotImplementedError(f"SharedNT does not support type: {nt.__class__.__name__}")
